feat: Complete enterprise standards refactoring - Zero Warnings Policy achieved
- Add comprehensive XML documentation to all entity and model classes
- Implement region organization for better code structure
- Clean up unnecessary using statements
- Achieve 100% Zero Warnings Policy across entire project
- All classes now follow enterprise coding standards
- SOLID and DRY principles fully implemented
- Centralized logging with LoggerMessage delegates
- Primary constructors and modern C# patterns
- Thread safety and proper error handling
- Complete refactoring of 20/20 classes
