+++ src/SmartRAG/Interfaces/IDocumentService.cs
@@ -1,7 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Threading.Tasks;
 using SmartRAG.Entities;
 
 namespace SmartRAG.Interfaces;
