+++ src/SmartRAG/Services/DocumentSearchService.cs
@@ -4,7 +4,7 @@ using Microsoft.Extensions.Options;
 using SmartRAG.Entities;
 using SmartRAG.Interfaces;
 using SmartRAG.Models;
-using SmartRAG.Services.Logging;
+using SmartRAG.Services;
 
 namespace SmartRAG.Services;
