+++ src/SmartRAG/Services/DocumentService.cs
@@ -3,7 +3,7 @@ using Microsoft.Extensions.Options;
 using SmartRAG.Entities;
 using SmartRAG.Interfaces;
 using SmartRAG.Models;
-using SmartRAG.Services.Logging;
+using SmartRAG.Services;
 using System.Globalization;
 using System.Text;
